+++ R/set_tracker_parameters.R
@@ -1,0 +1,29 @@
+#' Set the properties of the eye-tracker used in recording the data
+#'
+#' This function is used to specify the properties of the eye-tracker used to record the data. 
+#' This generates a list object within the workspace that is then accessible by other functions.
+#' If these are not set by the user, then default parameters are set and a warning is issued in every call to relevant functions
+#'
+#' @param sample_frequency the frequency with which samples are recorded by the eye-tracker
+#' @param viewing_distance_cm viewing distance in cm. Default of 60cm.
+#' @param screen_width_cm Default is 53 cm (standard 24" monitor).
+#' @param screen_height_cm Default is 30 cm (standard 24" monitor).
+#' @param screen_width_pixels used in conversion of pixel values. Default is 1920 pixels.
+#' @export screen_height_pixels used in conversion of pixel values. Default is 1920 pixels.
+#'
+#' @examples combine_eyes(HCL, method = "average")
+#'
+
+set_tracker_parameters <- function(sample_frequency = NULL, 
+                                   viewing_distance_cm = 60,
+                                   screen_width_cm = 53,
+                                   screen_height_cm = 30,
+                                   screen_width_pixels = 1920,
+                                   screen_height_pixels = 1080) {
+  
+  if 
+ 
+  
+  
+   
+}